Finish initial version
Everything's in place now for the end-to-end test to pass. Had to make sure to
exclude zero summands from the output, but everything else was fine.

There's definitely room for improvement in terms of edge case handling and
simplification. For instance, negative coefficients will probably work, but the
output wouldn't look great. Also, as mentioned in a previous commit message,
creating a presenter class for the human readable output could go a long way
towards simplifying the logic in Polynomial#to_s.